requirements.txt
numpy
Pillow
pyyaml
openai
regex
streamlit
